# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sorolens API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Sorolens indexes and surfaces data from Soroban smart contracts: events,
    invocations, storage, and per-contract analytics. This spec is the single
    source of truth for the auto-generated Go client (packages/go-client) and
    matches the live routes in apps/api/internal/router/router.go.

    Authentication: most read endpoints accept an optional API key via
    `Authorization: Bearer <token>` or `X-API-Key: <token>`. API-key management
    additionally requires an admin identity via role headers
    (`X-User-ID`, `X-GitHub-Login`, `X-User-Sub`, or `X-GitHub-ID`).
  contact:
    name: Sorolens
servers:
  - url: https://api.sorolens.xyz
    description: Production
...
paths:
  /api/v1/dlq:
    get:
      tags: [DLQ]
      summary: List dead-lettered indexer events
      description: Events that exhausted retries, newest-first, opaque cursor pagination.
      operationId: listFailedEvents
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
          description: Page size (default 50).
      responses:
        "200":
          description: Dead-lettered events
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/FailedEvent"
                  next_cursor:
                    type: string
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/dlq/{id}/requeue:
    post:
      tags: [DLQ]
      summary: Requeue a dead-lettered event
      description: Re-inserts the parked event into the events table and clears the DLQ row on success.
      operationId: requeueFailedEvent
      security:
        - ApiKeyAuth: [write:contracts]
          RoleAuth: [contributor]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The DLQ row's surrogate id.
      responses:
        "200":
          description: Requeued
          content:
            application/json:
              schema:
                type: object
                required: [id, event_id, status]
                properties:
                  id:
                    type: integer
                    format: int64
                  event_id:
                    type: string
                  status:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/stats/global:
    get:
      tags: [Stats]
      summary: Global index statistics
      operationId: getGlobalStats
      security:
        - ApiKeyAuth: [read:contracts]
      responses:
        "200":
          description: Global counts across all tracked contracts
          content:
            application/json:
              schema:
                type: object
                required: [tracked_contracts, total_events, total_invocations, total_storage_entries]
                properties:
                  tracked_contracts:
                    type: integer
                    format: int64
                  total_events:
                    type: integer
                    format: int64
                  total_invocations:
                    type: integer
                    format: int64
                  total_storage_entries:
                    type: integer
                    format: int64
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/events:
    get:
      tags: [Events]
      summary: List events across all tracked contracts
      operationId: listAllEvents
      description: Cross-contract events explorer feed, newest first.
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: contract_id
          in: query
          required: false
          schema:
            type: string
          description: Contract ID prefix (case-insensitive); a full ID matches one contract.
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [contract, system, diagnostic]
        - $ref: "#/components/parameters/NetworkParam"
        - name: since
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Inclusive lower bound on ledger_closed_at (RFC 3339).
        - name: until
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Inclusive upper bound on ledger_closed_at (RFC 3339).
      responses:
        "200":
          description: Page of events
          content:
            application/json:
              schema:
                type: object
                required: [events, next_cursor]
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
                  next_cursor:
                    type: string
                    description: Opaque cursor for the next (older) page; empty when no more pages.
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/compare:
    get:
      tags: [Stats]
      summary: Compare up to four contracts side by side
      operationId: compareContracts
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: string
          description: 1 to 4 comma-separated contract IDs.
        - name: window
          in: query
          required: false
          schema:
            type: string
            enum: [24h, 7d, 30d]
            default: 24h
      responses:
        "200":
          description: Comparison, one entry per requested contract in request order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompareResponse"
        "422":
          $ref: "#/components/responses/InvalidInput"
  /api/v1/contracts:
    get:
      tags: [Contracts]
      summary: List tracked contracts
      operationId: listContracts
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter by contract status (pending, active, backfilling, error, paused).
      responses:
        "200":
          description: Page of contracts
          headers:
            X-Cache:
              $ref: "#/components/headers/XCache"
          content:
            application/json:
              schema:
                type: object
                required: [contracts, next_cursor]
                properties:
                  contracts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Contract"
                  next_cursor:
                    type: string
                    description: Opaque cursor for the next page; empty when no more pages.
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Contracts]
      summary: Register a contract for tracking
      operationId: registerContract
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, network]
              properties:
                id:
                  type: string
                  pattern: "^C[A-Za-z0-9]{55}$"
                  description: Stellar contract ID (56 chars, starts with C).
                network:
                  type: string
                  enum: [testnet, mainnet, futurenet, standalone]
                label:
                  type: string
                  description: Optional human-friendly label.
      responses:
        "201":
          description: Contract registered (status pending until backfill completes).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}:
    get:
      tags: [Contracts]
      summary: Get one contract
      operationId: getContract
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: The contract
          headers:
            X-Cache:
              $ref: "#/components/headers/XCache"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/events:
    get:
      tags: [Events]
      summary: List contract events
      operationId: listContractEvents
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
        - name: type
          in: query
          required: false
          schema:
            type: string
          description: Event type filter.
        - name: from
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Lower ledger bound (inclusive).
        - name: to
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Upper ledger bound (inclusive).
      responses:
        "200":
          description: Page of events
          content:
            application/json:
              schema:
                type: object
                required: [events, next_cursor]
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
                  next_cursor:
                    type: string
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/events.csv:
    get:
      tags: [Events]
      summary: Export contract events as CSV
      description: >
        Streams every matching event as CSV instead of paginating them, so a
        large contract can be handed to a spreadsheet or an analyst in one
        request. Rows are ordered by ledger ascending, then event id, and the
        file is deterministic for an unchanged store.

        Unlike the JSON listing there is no cursor and no row cap: an export is
        meant to be complete. The same `type`, `from` and `to` filters apply,
        but `topic` and `in_successful_call` are not supported here.

        The header row is always present, so an unknown contract yields a
        file with no data rows rather than a 404. Free-text columns are
        prefixed with an apostrophe when they start with a character a
        spreadsheet would treat as a formula.
      operationId: exportContractEventsCSV
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/NetworkParam"
        - name: type
          in: query
          required: false
          schema:
            type: string
          description: Event type filter.
        - name: from
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Lower ledger bound (inclusive). Must not exceed `to`.
        - name: to
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Upper ledger bound (inclusive). Must not be below `from`.
      responses:
        "200":
          description: >
            CSV export of matching contract events, served as a downloadable
            attachment with `Content-Type: text/csv; charset=utf-8` and
            `Content-Disposition: attachment; filename="<id>-events.csv"`.
            The columns are id, contract_id, network, ledger,
            ledger_closed_at, tx_hash, type, topic_xdr, value_xdr,
            topic_decoded, value_decoded, in_successful_call.
          content:
            text/csv:
              schema:
                type: string
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/invocations:
    get:
      tags: [Invocations]
      summary: List contract invocations
      operationId: listContractInvocations
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [SUCCESS, FAILED, NOT_FOUND]
          description: Invocation status filter.
        - name: function_name
          in: query
          required: false
          schema:
            type: string
          description: Function name filter.
        - name: from
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Lower ledger bound (inclusive).
        - name: to
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Upper ledger bound (inclusive).
      responses:
        "200":
          description: Page of invocations
          content:
            application/json:
              schema:
                type: object
                required: [invocations, next_cursor]
                properties:
                  invocations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invocation"
                  next_cursor:
                    type: string
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/storage:
    get:
      tags: [Storage]
      summary: List contract storage entries
      operationId: listContractStorage
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
        - name: durability
          in: query
          required: false
          schema:
            type: string
            enum: [temporary, persistent, instance]
          description: Storage durability filter.
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [live, archived, deleted]
          description: Storage entry status filter.
      responses:
        "200":
          description: Page of storage entries
          content:
            application/json:
              schema:
                type: object
                required: [storage, next_cursor]
                properties:
                  storage:
                    type: array
                    items:
                      $ref: "#/components/schemas/StorageEntry"
                  next_cursor:
                    type: string
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/stats:
    get:
      tags: [Stats]
      summary: Per-contract statistics
      operationId: getContractStats
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - name: window
          in: query
          required: false
          schema:
            type: string
          description: Aggregation window. 7d and 30d map to 7/30 days; anything else is 24 hours.
      responses:
        "200":
          description: Contract statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractStats"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/forecast:
    get:
      tags: [Forecast]
      summary: Forecast contract usage over the next N days
      operationId: getContractForecast
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - name: horizon
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 365
          description: Number of days to forecast (default 30; optional trailing d accepted).
      responses:
        "200":
          description: Forecast series for fees, invocations, and events
          content:
            application/json:
              schema:
                type: object
                required: [contract_id, lookback_days, series]
                properties:
                  contract_id:
                    type: string
                  lookback_days:
                    type: integer
                  series:
                    type: array
                    items:
                      $ref: "#/components/schemas/ForecastSeries"
        "400":
          description: Invalid horizon parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/snapshot:
    get:
      tags: [Snapshots]
      summary: Snapshot of contract storage at a historical ledger
      operationId: getContractSnapshot
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - name: ledger
          in: query
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Ledger to snap (greater than zero).
      responses:
        "200":
          description: Storage snapshot plus the last event at or before the ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractSnapshot"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/snapshot.json:
    get:
      tags: [Snapshots]
      summary: Portable JSON export of a contract's current state
      operationId: exportContractSnapshot
      description: |
        Returns the contract's current state as a single versioned JSON
        document: metadata, storage, recent events, and summary. The shape is
        frozen by `schema_version` and is deterministic for an unchanged store
        (stable field order, sorted collections, no wall-clock fields), so
        exports can be archived or diffed. When the client advertises
        `Accept-Encoding: gzip` the body is gzip-compressed.
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: Versioned contract snapshot export
          headers:
            Content-Disposition:
              schema:
                type: string
              description: attachment; filename="<contract_id>-snapshot.json"
            Content-Encoding:
              schema:
                type: string
              description: gzip when the client advertises Accept-Encoding gzip.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractSnapshotExport"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/stream:
    get:
      tags: [Events]
      summary: Poll the most recent contract events
      operationId: streamContractEvents
      description: |
        Short-poll convenience used by the dashboard. Returns the 20 most
        recent events as JSON (not SSE).
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: Most recent events
          headers:
            Last-Modified:
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/contracts/{id}/summary:
    get:
      tags: [Stats]
      summary: Get a contract's dashboard summary
      operationId: getContractSummary
      description: Contract metadata, stats, latest event and invocation, and health score in one call.
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: The summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractSummary"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/stream/events:
    get:
      tags: [Events]
      summary: Stream new events (Server-Sent Events)
      operationId: streamAllEvents
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - name: contract_id
          in: query
          required: false
          schema:
            type: string
          description: Only stream events for this contract.
      responses:
        "200":
          description: An SSE stream; each `data:` line is a JSON Event.
          content:
            text/event-stream:
              schema:
                type: string
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/api-keys:
    get:
      tags: [API Keys]
      summary: List API keys
      operationId: listApiKeys
      security:
        - ApiKeyAuth: [admin:*]
        - RoleAuth: [admin]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Page of API keys
          content:
            application/json:
              schema:
                type: object
                required: [api_keys, next_cursor]
                properties:
                  api_keys:
                    type: array
                    items:
                      $ref: "#/components/schemas/APIKey"
                  next_cursor:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenScope"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [API Keys]
      summary: Create an API key
      operationId: createApiKey
      security:
        - ApiKeyAuth: [admin:*]
        - RoleAuth: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name:
                  type: string
                  description: Human-friendly key name.
                scopes:
                  type: array
                  items:
                    type: string
                    enum: [read:contracts, write:contracts, read:watchdog, admin:*]
                  minItems: 1
                  description: Scopes granted to the key.
      responses:
        "201":
          description: Key created; the plaintext key is returned once.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIKey"
                  - type: object
                    required: [key]
                    properties:
                      key:
                        type: string
                        description: Plaintext token, shown once.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenScope"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/api-keys/{id}:
    delete:
      tags: [API Keys]
      summary: Revoke an API key
      operationId: revokeApiKey
      security:
        - ApiKeyAuth: [admin:*]
        - RoleAuth: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: API key ID.
      responses:
        "204":
          description: Key revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenScope"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/watchdog/contracts/{id}:
    get:
      tags: [Watchdog]
      summary: Get one monitored contract
      operationId: getMonitoredContract
      security:
        - ApiKeyAuth: [read:watchdog]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: The monitored contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoredContract"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "Scoped API key. Send as `Authorization: Bearer <token>` (or `X-API-Key`)."
    RoleAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: |
        Role-based auth. Resolve a user identity from one of
        X-User-ID, X-GitHub-Login, X-User-Sub, or X-GitHub-ID and enforce the
        minimum role declared next to this security scheme.
    IdentityHeaderAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: Watchlist endpoints read the user ID from X-User-ID.
  parameters:
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Opaque pagination cursor from a previous response.
...